requirements.txt
streamlit
python-dotenv
google-generativeai
notion-client
requests
serpapi
